Change converters precedence and enrich readme (#119)
* Change converters precedence and enrich readme
* Fix test with floating point value and change readme
* Add math context for comparing doubles in test

+++ src/test/java/io/tarantool/driver/integration/TestComposite.java
@@ -8,5 +8,5 @@ public class TestComposite {
     public String field1;
     public Integer field2;
     public Boolean field3;
-    public Float field4;
+    public Double field4;
 }
